Restructure base and config into proper modules
- Create dedicated modules for base and config functionality
- Move enums and data_transfer_models into the base module
- Maintain backward compatibility with wrapper modules
- Improve code organization while ensuring no API breakage
- Add proper imports and exports in __init__.py files
